Moved PytestTestHelper into test/helper.py
